Add files via upload
As the project title suggests, this is a simple SSA (Standalone Single Application) , which breaks the given input which is supposed to be in English , and maps it to the word in French.

Technical Stack :
Data Structures used are : Dictionaries , Tuples & Lists

In order to run the App :

->Install the application using translator file provided along this text.

->open the folder where you have installed it to.

->open folder 'dist'

->and run the Eng2French.exe 

And that is it. 
please provide your feedback.
note:this is a application without any GUI; I am still working on 'tkinter' module of python.
once done i'll give a update to this.